Clean edit view code
