+++ packages/strapi-plugin-content-manager/admin/src/containers/EditView/init.js
@@ -1,45 +1,5 @@
-import { get } from 'lodash';
-
-const setDefaultForm = attributes => {
-  return Object.keys(attributes).reduce((acc, current) => {
-    const attribute = get(attributes, [current], {});
-    const {
-      default: defaultValue,
-      type,
-      required,
-      min,
-      repeatable,
-    } = attribute;
-
-    if (type === 'json') {
-      acc[current] = null;
-    }
-
-    if (defaultValue !== undefined) {
-      acc[current] = defaultValue;
-    }
-
-    if (type === 'group') {
-      if (required === true) {
-        acc[current] = repeatable === true ? [] : {};
-      }
-
-      if (min && repeatable === true) {
-        acc[current] = [];
-
-        for (let i = 0; i < min; i++) {
-          acc[current].push({ _temp__id: i });
-        }
-      }
-    }
-
-    return acc;
-  }, {});
-};
-
 function init(initialState) {
   return initialState;
 }
 
 export default init;
-export { setDefaultForm };
